Updating everything to compile.
Also Proposing some changes to schema naming, and adding a stub ga4gh.avdl.

What we really need to do is have one repo for all the GA4GH schemas that's flat and includes *just* the .avdl files, so they can be included as a submodule. That or invent much more sophisticated Avro dependency management.